requirements.txt
beautifulsoup4
lxml
discord.py
PyYAML
requests
